Fix: Fixed more edge cases with the value displays
Fixed another edge case with the price and percentage value displays, used recursion to fix some overflows caused by rounding, and added recursion and buffers so that values are not drawn until complete properties of value are known. E.g., 99.999 in a 4-character space would become 99.10% which is incorrect, now it becomes 100% as the rounding is detected.