Feature: New tool to change default coins (ESPIR_Assist)
I created a small Python tkinter program that makes the process of adding new default coins (with bitmap images) MUCH simpler.

The program takes the new coin ID, code, image of the logo, and the code of the coin to replace. The user selects the bitmap colour, logo colour, and accent colour. The user then selects the library and ticker directory, hits the 'REPLACE' button, and the files are modified automatically to add the new default coin. All the user needs to do is get the coin ID and the logo image, as well as know those directory's locations. Everything except getting the coingecko ID and uploading the sketch is now done on this application.