Feature: Added display fix functionality to ESPIR_Assist
If using this software with an ST7735 display, there are 2 main types of display you will come across, black tab and green tab. If your display is a green tab and you use the black tab settings or vice versa, colours invert and things become misaligned. All the functionality does is change the display initialisation to use the other type, and changes the colours in Colours.h from RGB 5:6:5 to BGR 5:6:5, this means the code can easily be changed using the software to work on both types of display with very minimal effort. Some changes have been made to the Colours.h to make sure it is clear what the current code is set up to run on (Green or black tab).