Fix: Crash when adding new coins
On the 2nd added coin (straight after the other one), once a certain limit was exceeded, a crash would occur. This crash was caused by undefined behaviour due to a missing terminating character. The string is now initialised as a char array of terminating characters to prevent this from happening. Also changed from freeing the string and reallocating to simply clearing it, this doesn't impact anything as inputs are copied anyway (not sure why I did it like that in the first place!).